requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
distro==1.9.0
elevenlabs==1.9.0
exceptiongroup==1.2.2
fireworks-ai==0.15.4
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
numpy==2.0.2
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
scipy==1.13.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
typing_extensions==4.12.2
urllib3==2.2.3
webrtcvad==2.0.10
websockets==13.1
